color: Add ColorCanvasExtensions runtime flag
These features being developed behind --enable-color-correct-rendering
and --enable-experimental-canvas-features.

We're turning on  --enable-color-correct-rendering for all layout tests,
some of which also use --enable-experimental-canvas-features, but do
not want to opt in to new canvas behaviors (at least yet).

To avoid this, use the Blink feature flag ColorCanvasExtensions. While
we're in the neighborhood, pull out the uses of the flag
--enable-color-correct-rendering-default-mode and related code.

BUG=724733

Review-Url: https://codereview.chromium.org/2893243003
Cr-Original-Commit-Position: refs/heads/master@{#474921}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1e300b0f88b6a65a87c9b27cd3e327499c25e86