Make CSSKeywordValue's value attribute mutable
Also update the attribute name to "value" to match the spec, and
change the type to a plain String, to avoid interning random
developer strings inside AtomicString when we don't need to.

In the original spec for Typed OM, every StyleValue subclass was
immutable, so you'd need to create new ones if you wanted to udpate
something. It turns out that this can be expensive due to garbage
collection churn, and it's now been changed so that StyleValue
subclasses are mutable.

CSSKeywordValue spec (note no "readonly" keyword on value now)
https://drafts.css-houdini.org/css-typed-om/#csskeywordvalue

BUG=545318

Review-Url: https://codereview.chromium.org/2893823002
Cr-Original-Commit-Position: refs/heads/master@{#474962}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2656da0429614ad0134c22e9695a619907d11a78