Allow storing multiple replacements on SpellCheckResult
I'm working on adding support for the Android spellcheck suggestion list that
appears in a native text box when you tap on a misspelled word. It appears that
I need to modify this piece of infrastructure to support passing multiple
suggestions through to the Spelling marker that gets added.

For reference, we construct the SpellCheckResult objects from the Android
spellchecker results in SpellCheckerSessionBridge::ProcessSpellCheckResults().

Alternatively, ContextMenuClientImpl has logic for calling the spellchecker
again to get the list of suggestions that we might be able to use:
https://chromium.googlesource.com/chromium/src/+/adc8910776c64a876f0422454da618a37d01167b/third_party/WebKit/Source/web/ContextMenuClientImpl.cpp#346

However, I talked to aelias and he said it would be better to store the
suggestions the first time we call the spellchecker, at the time we add the
marker, to avoid going back and forth unnecessarily.

BUG=715365

Review-Url: https://codereview.chromium.org/2848943002
Cr-Original-Commit-Position: refs/heads/master@{#475058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c45f50742410a644a054325478d885e61d286713