Rename RenderedDocumentMarker to TextMatchMarker
RenderedDocumentMarker used to be created for all DocumentMarker types. Now it's
only created for TextMatchMarkers. Renaming it to TextMatchMarker sets us up to
move all of the TextMatch-specific functionality out of DocumentMarker into this
class. I eventually want to do this for all the other marker types as well.

BUG=707867

Review-Url: https://codereview.chromium.org/2898953005
Cr-Original-Commit-Position: refs/heads/master@{#475070}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 802826d9bacfad322a4ec415373d1ed3f317a2a2