Add member TextIterator::text_node_
This patch adds a new member TextIterator::text_node_ to record
the text node that is being handled in TextIterator. As a result,
text node handling and the remaining part of TextIterator share one
less member, so that the encapsulation of text node handling is
easier.

BUG=721957
TEST=n/a; no behavioral change

Review-Url: https://codereview.chromium.org/2902173003
Cr-Original-Commit-Position: refs/heads/master@{#475237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b109b249a23d3146575b12b1fa4781ade53dac28