Make UseCounter take a LocaFrame instead of any Frame
It doesn't make sense for a UseCounter to be triggered by a RemoteFrame.
Update all the callers to provide a LocalFrame (or in a couple cases
where we can't be sure to have one, skip reporting usage).

In a future patch I'd like to start depending on the fact that we know
the correct LocalFrame for a UseCounter invocation.

BUG=716565

Review-Url: https://codereview.chromium.org/2877893002
Cr-Original-Commit-Position: refs/heads/master@{#475268}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64819bf38f1754e6677f2be17335a2b8f8103060