Move many more classes to use WebLocalFrameBase over WebLocalFrameImpl.
In the majority of these case, we just adjust the header to include
core/frame/WebLocalFrameBase.h instead of wevb/WebLocalFrameImpl.h, as the
class is just needs to update the definition.

In some cases we also needed to change the downcaststs to WebLocalFrameBase.

BUG=708879

Review-Url: https://codereview.chromium.org/2900593002
Cr-Original-Commit-Position: refs/heads/master@{#475280}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd4e07b5898a1a7f287085aa900f0ffbb77acbe2