Disable style sharing outside tests.
This disables style sharing for most users in Canary, but leaves the code and
leaves it enabled in the layout tests. Once this bakes for a little while we
can then remove the code and update the tests.

Per the style-dev@ discussion and design doc:
https://groups.google.com/a/chromium.org/forum/#!topic/style-dev/8BV0QYubjos

Pateched in from esprehn's patch: http://crrev.com/2858423002

BUG=721517

Review-Url: https://codereview.chromium.org/2887593002
Cr-Original-Commit-Position: refs/heads/master@{#475317}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc275c99719dca651110ad31085c6401de79056a