Upstream service wrkr "window client" test to WPT
Re-locate test files to Web Platform Test directory for eventual
automated upstreaming.

The original version of this test registered a handler for the service
worker's "install" event after initial evaluation. Because this practice
is discouraged by the specification, it should be avoided. Re-factor the
test to immediately register the event handler and to preserve the
necessary behavior through an explicit message channel with the client.

Add a "use strict" directive to script bodies. Introduce a variable
declaration for a previously-undeclared binding. Correct a typo in an
error message. Ensure that each sub-test's asynchronous "cleanup" logic
is complete prior to advancing to the next sub-test.

BUG=688116
R=falken@chromium.org

Review-Url: https://codereview.chromium.org/2906963002
Cr-Original-Commit-Position: refs/heads/master@{#475409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d677eb1d7292c42d3902f28e2add76e9c65684d