Introduce a wrapper class to handle text node in TextIterator
This patch adds (in-place) a wrapper class for the text node handling code
in TextIterator. With this patch, TextIterator doesn't need to know the
details of the text layout, so it can switch to Layout NG easier.

A follow-up patch will remove the wrapper class to dedicated files.

BUG=721957
TEST=n/a; no behavioral changes

Review-Url: https://codereview.chromium.org/2903693005
Cr-Original-Commit-Position: refs/heads/master@{#475411}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8d639e9e2ac1811b38e6bacc13e73049b27750f4