color: Perform color transform in ImageBufferSurface
When we draw a canvas element to a AcceleratedImageBufferSurface or
to a UnacceleratedImageBufferSurface, we must transform the inputs to
the color space specified by the CanvasColorParams it was initialized
with.

To do this, make the SkiaPaintCanvas wrap the SkCanvas in a
SkColorSpaceXformCanvas.

BUG=725372
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2904143002
Cr-Original-Commit-Position: refs/heads/master@{#475412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 30a7f83076be71351b71dfcb6cefee6ec2a9faca