Retain is_current_color information in BorderValue constructor
This CL fixes the BorderValue constructor to retain the is_current_color
information which it was losing due to us resolving the StyleColor to
Color but still using it as a StyleColor. This meant we didn't have
the is_current_color information.

This fixes a bug in the code where we are losing information when
constructing BorderValue.

A unit test is also added to the effect.

BUG=628043

Review-Url: https://codereview.chromium.org/2913663002
Cr-Original-Commit-Position: refs/heads/master@{#475433}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de1b3604290f911cf4fc76988ee4d5b11a4f07f9