[LayoutNG] Use logical left of the line box to apply text-align
This patch changes text-align to use the logical left of the line box to
apply the CSS text-align property. This is set to the offset of
NGPhysicalLineBoxFragment, which is then copiped to
RootInlineBox::LogicalLeft.

This way is more efficient to move, the same as what the legacy does,
and better for painting since the line box does not have unused spaces.

BUG=636993
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_layout_ng

Review-Url: https://codereview.chromium.org/2906233002
Cr-Original-Commit-Position: refs/heads/master@{#475728}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1680fcaff1aa1c755f52a2b8942cf871b773a32