Generate enum/getters/setters/mappings for ruby-position.
Currently, the ruby-position property is a 'storage_only' field,
so it has no generated public getters/setters. This patch changes it to
a 'keyword' field so that its getters/setters can be generated as well.
We also generate the ERubyPosition enum and use the generated
CSSValueID <-> ERubyPosition mappings.

BUG=628043

Review-Url: https://codereview.chromium.org/2915683002
Cr-Original-Commit-Position: refs/heads/master@{#475817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e4c61cfefb77952159dae7efab12914e9f4e56f