Generate enum/getters/setters/mappings for line-break.
Currently, the line-break property is a 'storage_only' field,
so it has no generated public getters/setters. This patch changes it to
a 'keyword' field so that its getters/setters can be generated as well.
We also generate the ELineBreak enum and use the generated
CSSValueID <-> ELineBreak mappings.

BUG=628043

Review-Url: https://codereview.chromium.org/2911383002
Cr-Original-Commit-Position: refs/heads/master@{#475830}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51bbe33ffd7411539815a5f7e554b39b5e26eefd