Unified Autoplay: implement document user gesture restriction in Blink.
This is adding a new AutoplayPolicy in Blink that checks if the document
received a user gesture. It applies to <audio> and <video>.

Presubmit is ignored because pointerActionSequence does not support
keyboard events yet and this is required for the test.

BUG=715049
NOPRESUBMIT=true
R=dcheng@chromium.org, mustaq@chromium.org

Change-Id: Ie20d44d90fe8850c0e59aad5bcbef7ca4d2768d3
Reviewed-on: https://chromium-review.googlesource.com/507747
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#475960}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6850ab7401a3124a61684aaabb063e8dabb23ee