Generate enum/getters/setters/mappings for user-select.
Currently, the user-select property is a 'storage_only' field,
so it has no generated public getters/setters. This patch changes it to
a 'keyword' field so that its getters/setters can be generated as well.
We also generate the EUserSelect enum and use the generated
CSSValueID <-> EUserSelect mappings.

Diff of generated files:
https://gist.github.com/darrnshn/55f9abda5310dd37a0b60386f23e0e4a/revisions

BUG=628043

Review-Url: https://codereview.chromium.org/2911963004
Cr-Original-Commit-Position: refs/heads/master@{#475867}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: adccf92c92a9fa6094191e3938b3bd42948751fd