Make CounterDirectiveMap a class and implement a Clone() method.
CounterDirectiveMap is currently a typedef of a HashMap, which means
it can't be extended with the Clone method that is required by
StyleRareNonInheritedData when it becomes generated. This patch
makes it into a class and adds a Clone method so that it can be
copied by a generated StyleRareNonInheritedData.

BUG=628043

Review-Url: https://codereview.chromium.org/2907063002
Cr-Original-Commit-Position: refs/heads/master@{#475884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09c379de358b916213f0591a5214b60e76dfb224