Replace call sites to BorderValue functions to save the BorderValue construction cost
Currently we construct BorderValues via functions
ComputedStyle::Border{Left,Right,Start,End} only to access two methods
at most (Style and Width) or perform comparison. This patch splits out
the accessors for Style and Width and adds comparison functions to
ComputedStyle to save this construction cost.

We haven't removed the callsites in Layout land that return a
BorderValue and are saving it for a future patch to reduce complexity.

BUG=628043, 721287

Review-Url: https://codereview.chromium.org/2906253003
Cr-Original-Commit-Position: refs/heads/master@{#475896}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a4d00a687e0cf0ea029a68b2dd44f5814d539d0