Use remembered progress to handle replaced elements in TextIterator
This patch utilizes the "remembered progress" mechanism in TextIterator
to handle a replaced element when it emits multiple text runs. In this
way, the control flow becomes cleaner as we don't need to go into the
mail loop of TextIterator::Advance() again.

Follow-up patches will remove the return value of HandlePlacedElement()
because this patch makes it almost useless.

BUG=721957
TEST=n/a; shouldn't have any behavioral change

Review-Url: https://codereview.chromium.org/2912283002
Cr-Original-Commit-Position: refs/heads/master@{#475977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 529b9affa165a95a6b9369809ab26d7c5d37466b