Track slow paths in DisplayItemList
Previously, DisplayItemList only stored a boolean for whether or not
the entire DisplayItemList was suitable for gpu raster.  (This is
used for turning on msaa or not, but it hasn't been renamed).

PaintOpBuffer stores indirections to DisplayItemList and so needs to
properly track the number of slow paths stored inside of it.  The
solution is to push down the msaa veto logic from PaintController
down into PictureLayer and store slow path count the whole way.

I think this is more compatible with spv2, which will want to
combine paint artifacts and then decide if the result needs
msaa or not.

Also, issue 513016 is quite solved, so comments related to it can be
removed.
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel;master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2876033005
Cr-Original-Commit-Position: refs/heads/master@{#476040}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36ee28f4f311d6f307de81dc4fa42fed14c4bac7