Generate enum/getters/setters/mappings for user-modify.
Currently, the user-modify property is a 'storage_only' field,
so it has no generated public getters/setters. This patch changes it to
a 'keyword' field so that its getters/setters can be generated as well.
We also generate the EUserModify enum and use the generated
CSSValueID <-> EUserModify mappings.

Diff of generated files:
https://gist.github.com/21e3b6343b06a190a48ab3e22c9e10bc/revisions

BUG=628043

Review-Url: https://codereview.chromium.org/2912923003
Cr-Original-Commit-Position: refs/heads/master@{#476141}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4e5b1d42138414abcef8816f6af08fbebdbccd2