Worklet: Provide WorkerFetchContext for threaded worklets
This is a preparation CL for module loading for threaded worklets that is going
to be implemented on the off-main-thread-fetch feature.

This change provides WorkerFetchContext that is necessary for the
off-main-thread-fetch to worklet global scopes. The context is created on the
main thread before starting a worklet global scope and passed to worklet's
thread as WorkerThreadStartupData.

Note that this CL does not change the current script loading mechanism for
worklets yet.

Bug: 727194
Change-Id: I695bf8f64a583ac845a2d0a13076fda0667d8604
Reviewed-on: https://chromium-review.googlesource.com/519022
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#476246}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7d3d2fee88d26f0e6cb0fd224db028718813eb2