Do not use |unsigned char| as a hash key for blink::CollectionType enum.
This CL adds enum support to HashTraits and DefaultHash.

NodeListNodeData used the following HashMap to represent a map from a pair of
CollectionType and AtomicString to a node list.

    HeapHashMap<std::pair<unsigned char, AtomicString>, WeakMember<LiveNodeListBase>, ...>

This was not good for code readability.  Also, this code worked accidentally
because 0 CollectionType is valid but it's not used for this HashMap.

Change-Id: I0631c3b4f65f70a0ed445494e0e868215ff2c4e7
Reviewed-on: https://chromium-review.googlesource.com/517527
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yuta Kitamura <yutak@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#476572}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d5cfc6525d4701b6b478d04a09bb429293f600b5