XHR: Remove Blink.XHR.setRequestHeader.HeaderValueCategoryInRFC7230 counter
As discussed in
https://groups.google.com/a/chromium.org/d/msg/blink-dev/K5jd8Y5l6pI/KQW1nyAhBwAJ,
it should be safe to remove the use counter at this point and start
normalizing header values as mandated by the XHR and Fetch specs.

In this CL, we only drop the use counter and related code, thus greatly
simplifying SetRequestHeaderInternal(). The only reason that method was not
removed altogether in the same CL is that other methods call it to set the
Content-Type header without having to go throuh setRequestHeader() and
passing an ExceptionState to it.

There should be no behavior change with this CL.

Bug: 681769
Change-Id: I248e8085443e0dbaf55964eb618ceaa951de4782
Reviewed-on: https://chromium-review.googlesource.com/521105
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Takeshi Yoshino <tyoshino@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Raphael Kubo da Costa (rakuco) <raphael.kubo.da.costa@intel.com>
Cr-Original-Commit-Position: refs/heads/master@{#476702}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d562e29584e78ee12ff1295f6ca8878ca8544df6