Add SpellingMarker and GrammarMarker (subclasses of DocumentMarker)
In https://codereview.chromium.org/2904093002, I'm eliminating the
DocumentMarkerDetails subclass for TextMatchMarker in favor of storing the
TextMatchMarker-specific data directly on TextMatchMarker. In
https://codereview.chromium.org/2908643002, I'm creaing a similar subclass for
CompositionMarker. In this CL, I'm doing the same for SpellingMarker and
GrammarMarker, which means we can now eliminate DocumentMarkerDetails.

In a follow-up CL, I will make MarkerType() virtual so we won't have to store
the MarkerType as a field in DocumentMarker.

BUG=707867

Review-Url: https://codereview.chromium.org/2911723002
Cr-Original-Commit-Position: refs/heads/master@{#476419}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45ab0de82123f81a41e7af931ba30e180532066f