Move TextMatchMarker method implementations to TextMatchMarker.cpp
TextMatchMarker used to be RenderedDocumentMarker, which didn't have a .cpp
file, and had all its method implementations in a header file. This CL moves
TextMatchMarker's method implementations to its .cpp file for cleanliness, as
well as consistency with the classes for the other marker types.

BUG=707867

Review-Url: https://codereview.chromium.org/2919603005
Cr-Original-Commit-Position: refs/heads/master@{#476550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb2881f8de65ec1c7d64d92ce414a5f108a86270