Make PaintImages a class and implement a Clone() method.
Currently we use Vector<Persistent<StyleImage>> to represent a list
of paint images. When we generate StyleRareNonInheritedData, we need
this type to have a Clone method because it is stored inside a
std::unique_ptr. This patch makes this type its own class so that
it can be extended with a Clone method.

BUG=628043

Review-Url: https://codereview.chromium.org/2911343002
Cr-Original-Commit-Position: refs/heads/master@{#476580}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e887928f5fc67b86d3d2bf951c3022caaa2314e