Put TextIteratorTextNodeHandler and TextIteratorTextState on heap
This patch puts TextIteratorTextNodeHandle and TextIteratorTextState on heap,
so that we can use forward declaration, and also opens up possibility for
switching them to a Layout NG version.

Note: we can't only put TextIteratorTextNodeHandler on heap because it holds
a reference to TextIteratorTextState, but an on-heap object cannot hold reference
to a stack-allocated object.

BUG=721957
TEST=n/a; no behavioral change

Review-Url: https://codereview.chromium.org/2915043004
Cr-Original-Commit-Position: refs/heads/master@{#476605}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7fee9a6165a732a2cb346c9db63ab65adab5398